add: custom cart | default cart | and calculate every cart sum

+++ src/main/java/com/seminar/easyCookWeb/model/recipe/RecipeModel.java
@@ -41,7 +41,7 @@ public class RecipeModel {
 
     @ApiModelProperty(value = "食譜價格", example = "10.5")
     @Builder.Default
-    private int price = 0;
+    private double price = 0;
 
 
 //    @Builder.Default
